test(twitch): align live lifecycle proof (#81853) (thanks @edenfunf)
